Code reorg to take better advantage of cxx
Making use of the cxx binding generator is probably the best way to get rid of
the boilerplate, while keeping the rich integration with snort’s C++ types.

The code now expects:
 - snort to be present in /opt/snort (packages distributed through internal deb
   server for now, will be made public in the future).
 - a symlink to be created for include to snort’s include directory
 - cxxbridge (from cxxbridge-cmd) on the machine
 - C++ & Rust compile chain, including asan (by default).

The new makefile under “network_mapping” should result in creating the so file –
but I could not make it work with snort in time; looking forward to feedback.

The layout is not final (I think things are still too nested).